requirements.txt
streamlit==1.42.2
langchain==0.3.19
langchain-openai==0.3.7
langgraph==0.3.2
openai==1.65.2
python-dotenv==1.0.1
wikipedia==1.4.0
beautifulsoup4==4.13.3
requests==2.32.3
faiss-cpu==1.10.0
tiktoken==0.9.0
langchain-community==0.3.18
sseclient==0.0.27
fastapi==0.115.11
uvicorn==0.34.0
matplotlib==3.10.1
langfuse==2.59.7
networkx==3.4.2
duckduckgo_search==7.5.4
pydantic==2.8.2
chromadb
networkx
SQLAlchemy>=2.0
jinja2
pyrfc

